[#44] refactor: Member package 구조 변경
