+++ src/main/java/com/woowacourse/zzazanstagram/model/article/service/ArticleAssembler.java
@@ -1,11 +1,11 @@
-package com.woowacourse.zzazanstagram.model.article;
+package com.woowacourse.zzazanstagram.model.article.service;
 
 import com.woowacourse.zzazanstagram.model.article.domain.Article;
 import com.woowacourse.zzazanstagram.model.article.domain.vo.Contents;
 import com.woowacourse.zzazanstagram.model.article.domain.vo.Image;
 import com.woowacourse.zzazanstagram.model.article.dto.ArticleRequest;
 import com.woowacourse.zzazanstagram.model.article.dto.ArticleResponse;
-import com.woowacourse.zzazanstagram.model.member.Member;
+import com.woowacourse.zzazanstagram.model.member.domain.Member;
 
 import java.time.LocalDateTime;
