+++ src/main/java/com/woowacourse/zzazanstagram/model/article/service/ArticleService.java
@@ -1,12 +1,11 @@
 package com.woowacourse.zzazanstagram.model.article.service;
 
-import com.woowacourse.zzazanstagram.model.article.ArticleAssembler;
 import com.woowacourse.zzazanstagram.model.article.domain.Article;
 import com.woowacourse.zzazanstagram.model.article.dto.ArticleRequest;
 import com.woowacourse.zzazanstagram.model.article.dto.ArticleResponse;
 import com.woowacourse.zzazanstagram.model.article.repository.ArticleRepository;
-import com.woowacourse.zzazanstagram.model.member.Member;
-import com.woowacourse.zzazanstagram.model.member.MemberService;
+import com.woowacourse.zzazanstagram.model.member.domain.Member;
+import com.woowacourse.zzazanstagram.model.member.service.MemberService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
