+++ src/main/java/com/woowacourse/zzazanstagram/model/member/dto/MemberLoginRequest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.zzazanstagram.model.member;
+package com.woowacourse.zzazanstagram.model.member.dto;
 
 public class MemberLoginRequest {
     private String email;
