+++ src/main/java/com/woowacourse/zzazanstagram/model/member/repository/MemberRepository.java
@@ -1,7 +1,8 @@
-package com.woowacourse.zzazanstagram.model.member;
+package com.woowacourse.zzazanstagram.model.member.repository;
 
-import com.woowacourse.zzazanstagram.model.member.vo.Email;
-import com.woowacourse.zzazanstagram.model.member.vo.NickName;
+import com.woowacourse.zzazanstagram.model.member.domain.Member;
+import com.woowacourse.zzazanstagram.model.member.domain.vo.Email;
+import com.woowacourse.zzazanstagram.model.member.domain.vo.NickName;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
